ArgillaforDataAnnotationofOdiaDataset_SamiritSaha
Hi there, I'm Samirit Saha, an intern at OdiaGenAI and the writer of this code. This is the code which allows the user to upload a dataset and then evaluate the dataset (on a score of 1 to 5) and annotate it. For this code, we make use of the dolly-odia-15k dataset from OdiaGenAI. With this code, this dataset has been uploaded to the following argilla workspace where it can be evaluated and annotated: https://samirit-dataannotation.hf.space/datasets. Access the given url to evaluate and annotate the dataset. Also, if instead of the dataset that is being annotated in the given code, you want to annotate some other dataset, in the following line of the code: hf_dataseta = load_dataset('OdiaGenAI/dolly-odia-15k', split='train'), just change 'OdiaGenAI/dolly-odia-15k' with the address of your code in Huggingface. The username is 'argilla' and password is '12345678'  for the argilla workspace.

Signed-off-by: Samirit <73527396+Samirit@users.noreply.github.com>