Python code for Argilla UI for Data Annotation of Odia Dataset by intern Samirit Saha for OdiaGenAI (#20)
This is the code which allows the user to upload a dataset and then evaluate the dataset (on a score of 1 to 5) and annotate it. For this code, we make use of the dolly-odia-15k dataset from OdiaGenAI. With this code, this dataset has been uploaded to the following argilla space where it can be annotated: https://samirit-dataannotation.hf.space/datasets. Access the given url to evaluate and annotate the dataset. Also, if instead of the dataset that is being annotated in the given code, you want to annotate some other dataset, in the following line of the code: hf_dataseta = load_dataset('OdiaGenAI/dolly-odia-15k', split='train'), just change 'OdiaGenAI/dolly-odia-15k' with the address of your code in Huggingface.

Signed-off-by: Samirit <73527396+Samirit@users.noreply.github.com>